Fix broken graphs
Turns out ab returns it's data in gnuplot sorted by the total time
it takes for the request to finish, so we wind up with a hockey
stick always.

Update to do some sorting in subshell in the plot command.  The resulting
graph is a bit harder to make sense of - may need to fix this up further